+++ data-analytics/dataflow-python-examples/dataflow_python_examples/data_lake_to_mart_cogroupbykey.py
