+++ data-analytics/dataflow-python-examples/dataflow_python_examples/resources/orders_denormalized.json
